+++ tests/smoke/CSharp50/AsyncAwait/AwaitTaskDelayInsteadOfCallingThreadSleep/TaskDelayCanBeAwaitedInsteadOfCallingThreadSleep.cs
@@ -1,0 +1,18 @@
+﻿// ReSharper disable All
+
+// Expected number of suggestions: 2
+
+using System.Threading;
+using static System.Threading.Thread;
+
+namespace CSharp50.AsyncAwait.AwaitTaskDelayInsteadOfCallingThreadSleep
+{
+    class TaskDelayCanBeAwaitedInsteadOfCallingThreadSleep
+    {
+        async void CanBeAwaited()
+        {
+            Thread.Sleep(0);
+            Sleep(0);
+        }
+    }
+}