+++ tests/smoke/CSharp50/AsyncAwait/AwaitTaskInsteadOfCallingTaskWait/TaskCannotBeAwaitedInsteadOfCallingTaskWait.cs
@@ -1,0 +1,15 @@
+﻿// ReSharper disable All
+
+using System.Threading.Tasks;
+
+namespace CSharp50.AsyncAwait.AwaitTaskInsteadOfCallingTaskWait
+{
+    class TaskCannotBeAwaitedInsteadOfCallingTaskWait
+    {
+        void CannotBeAwaited()
+        {
+            var task = new Task(() => { });
+            task.Wait(0);
+        }
+    }
+}